Update dependencies: remove typescript dependency and require version 1.4.1 of typescript-compiler.
The adapter can work with any 1.x version of typescript-compiler, but the package.json
requires version 1.4.1 explicitely. This is the only way I managed to get it to work
properly as both "1.x" and "^1.4" result in npm installing version 1.4.1-2, which does
not match semver "1.x" (call to semver.satisfies(...) fails), as whatever comes after
the hyphen indicates a pre-release version (alpha, rc...).