Use semver in devDependencies: `latest` -> `*`
The `latest` version is simply the last version that was published to NPM. For example, this could be a pre-release version or I believe it could even be a bugfix to an old version of the module.

I think `*` is probably what is wanted as it will install the latest stable release according to semver. You could also use `>= 0`, which does the same thing.